Refactor editor gizmo into focused modules
Split monolithic 1005-line index.ts into 9 focused modules:
camera, discovery, drag, gizmo, indicator, input, math-utils,
persistence, and selection. Index.ts is now an 82-line orchestrator.
