+++ extensions/dcl-editor-save.ts
@@ -1,13 +1,9 @@
 /**
  * DCL Editor Save Extension
  *
- * Provides `/save-editor` command to apply in-scene gizmo changes to source code.
- * Also prompts on session start and blocks deploy if pending changes exist.
- *
- * Flow:
- * 1. mv editor-scene.json → editor-scene.json.bkp (atomic, no data loss)
- * 2. Agent reads .bkp and patches Transform calls in source code
- * 3. On success → agent deletes .bkp
+ * Provides `/save-editor` command and startup prompt for pending editor changes.
+ * The actual patching logic is in the editor-gizmo skill — this extension just
+ * detects changes, creates the backup, and tells the agent to apply them.
  */
 
 import type { ExtensionFactory } from "@mariozechner/pi-coding-agent";
@@ -17,89 +13,6 @@ import { fileExists, findSceneRoot } from "./scene-utils.js";
 
 const EDITOR_FILE = "src/__editor/editor-scene.json";
 const BACKUP_FILE = "src/__editor/editor-scene.json.bkp";
-const RAD_TO_DEG = 180 / Math.PI;
-
-interface Vec3 {
-  x: number;
-  y: number;
-  z: number;
-}
-
-interface Quaternion extends Vec3 {
-  w: number;
-}
-
-interface TransformData {
-  position?: Vec3;
-  rotation?: Quaternion;
-  scale?: Vec3;
-}
-
-interface EntityChange {
-  components: {
-    Transform?: TransformData;
-  };
-}
-
-/** Convert quaternion to euler angles (degrees) for human-readable code */
-function quatToEuler(q: Quaternion): Vec3 {
-  const sinRoll = 2 * (q.w * q.x + q.y * q.z);
-  const cosRoll = 1 - 2 * (q.x * q.x + q.y * q.y);
-  const x = Math.atan2(sinRoll, cosRoll) * RAD_TO_DEG;
-
-  const sinPitch = 2 * (q.w * q.y - q.z * q.x);
-  const y =
-    Math.abs(sinPitch) >= 1
-      ? Math.sign(sinPitch) * 90
-      : Math.asin(sinPitch) * RAD_TO_DEG;
-
-  const sinYaw = 2 * (q.w * q.z + q.x * q.y);
-  const cosYaw = 1 - 2 * (q.y * q.y + q.z * q.z);
-  const z = Math.atan2(sinYaw, cosYaw) * RAD_TO_DEG;
-
-  return { x: round(x), y: round(y), z: round(z) };
-}
-
-function round(v: number): number {
-  return Math.round(v * 100) / 100;
-}
-
-function formatRotation(rotation: Quaternion): string | null {
-  const euler = quatToEuler(rotation);
-  if (euler.x === 0 && euler.y === 0 && euler.z === 0) return null;
-  return `Quaternion.fromEulerDegrees(${euler.x}, ${euler.y}, ${euler.z})`;
-}
-
-function buildAgentMessage(
-  entities: Record<string, EntityChange>
-): string {
-  const entries = Object.entries(entities);
-  const lines = entries.map(([name, entry]) => {
-    const t = entry.components?.Transform;
-    if (!t) return `- Entity "${name}": no transform changes`;
-
-    const pos = t.position
-      ? `Vector3.create(${t.position.x}, ${t.position.y}, ${t.position.z})`
-      : null;
-    const rot = t.rotation ? formatRotation(t.rotation) : null;
-
-    const parts = [];
-    if (pos) parts.push(`position = ${pos}`);
-    if (rot) parts.push(`rotation = ${rot}`);
-    return `- Entity "${name}": ${parts.join(", ")}`;
-  });
-
-  return [
-    `The user moved/rotated ${entries.length} object(s) in the scene editor.`,
-    `Update the Transform components in the scene source code to match these new values.`,
-    `Find each entity by its Name component value and update the position and rotation in the corresponding Transform.create() call.`,
-    `If the rotation is all zeros, you can omit the rotation field.`,
-    `After applying all changes, delete the backup file: src/__editor/editor-scene.json.bkp`,
-    ``,
-    `Changes:`,
-    ...lines,
-  ].join("\n");
-}
 
 /** Check if there are pending editor changes (either main file or interrupted backup) */
 async function findPendingChanges(
@@ -128,18 +41,14 @@ async function findPendingChanges(
   return null;
 }
 
-/** Read and parse the changes file, return entity count and agent message */
-async function readChanges(
-  filePath: string
-): Promise<{ count: number; message: string } | null> {
+/** Count entities in a changes file */
+async function countChanges(filePath: string): Promise<number> {
   try {
     const content = await readFile(filePath, "utf-8");
-    const data = JSON.parse(content) as Record<string, EntityChange>;
-    const count = Object.keys(data).length;
-    if (count === 0) return null;
-    return { count, message: buildAgentMessage(data) };
+    const data = JSON.parse(content);
+    return Object.keys(data).length;
   } catch {
-    return null;
+    return 0;
   }
 }
 
@@ -157,11 +66,16 @@ async function createBackup(sceneRoot: string): Promise<boolean> {
 }
 
 const extension: ExtensionFactory = (pi) => {
-  function sendEditorChanges(changes: { count: number; message: string }): void {
+  function sendApplyRequest(count: number): void {
     pi.sendMessage(
       {
         customType: "editor-save",
-        content: `Editor: ${changes.count} entity transform(s) to apply\n\n${changes.message}`,
+        content: [
+          `The user has ${count} pending editor change(s) to apply.`,
+          `Read the editor-gizmo skill for the full apply process.`,
+          `The changes are in: src/__editor/editor-scene.json.bkp`,
+          `Read that file, then follow the "Applying Editor Changes to Source Code" section of the skill.`,
+        ].join("\n"),
         display: true,
       },
       { triggerTurn: true, deliverAs: "nextTurn" }
@@ -184,8 +98,8 @@ const extension: ExtensionFactory = (pi) => {
         return;
       }
 
-      const changes = await readChanges(pending.path);
-      if (!changes) {
+      const count = await countChanges(pending.path);
+      if (count === 0) {
         ctx.ui.notify("Editor changes file is empty or invalid.", "warning");
         return;
       }
@@ -199,15 +113,14 @@ const extension: ExtensionFactory = (pi) => {
       }
 
       ctx.ui.notify(
-        `Applying ${changes.count} editor change(s) to source code...`,
+        `Applying ${count} editor change(s) to source code...`,
         "info"
       );
-      sendEditorChanges(changes);
+      sendApplyRequest(count);
     },
   });
 
-  // Use before_agent_start instead of session_start so the terminal's
-  // input handling is fully initialized (arrow keys work in the prompt).
+  // Prompt on session start if pending changes exist
   let editorPromptShown = false;
   pi.on("before_agent_start", async (_event, ctx) => {
     if (editorPromptShown) return;
@@ -219,12 +132,12 @@ const extension: ExtensionFactory = (pi) => {
     const pending = await findPendingChanges(sceneRoot);
     if (!pending) return;
 
-    const changes = await readChanges(pending.path);
-    if (!changes) return;
+    const count = await countChanges(pending.path);
+    if (count === 0) return;
 
     const label = pending.isBackup
-      ? `Found ${changes.count} interrupted editor change(s) from a previous session.`
-      : `Found ${changes.count} pending editor change(s).`;
+      ? `Found ${count} interrupted editor change(s) from a previous session.`
+      : `Found ${count} pending editor change(s).`;
 
     const apply = await ctx.ui.confirm(
       "Editor Changes",
@@ -235,7 +148,7 @@ const extension: ExtensionFactory = (pi) => {
     if (!pending.isBackup) {
       await createBackup(sceneRoot);
     }
-    sendEditorChanges(changes);
+    sendApplyRequest(count);
   });
 };
 
@@ -254,6 +167,5 @@ export async function getPendingEditorChanges(
   const pending = await findPendingChanges(sceneRoot);
   if (!pending) return 0;
 
-  const changes = await readChanges(pending.path);
-  return changes?.count ?? 0;
+  return await countChanges(pending.path);
 }
