+++ package.json
@@ -44,7 +44,7 @@
   "author": "Decentraland",
   "license": "Apache-2.0",
   "dependencies": {
-    "@mariozechner/pi-coding-agent": "^0.54.0",
+    "@mariozechner/pi-coding-agent": "^0.55.4",
     "playwright-core": "^1.58.2"
   },
   "devDependencies": {
