+++ skills/editor-gizmo/.gitignore
@@ -5,3 +5,4 @@
 !src/
 !src/__editor/
 !src/__editor/**
+src/__editor/editor-scene.json
